fix: Modal 에러 처리, 좋아요 버튼 실시간 반영되게 수정
