+++ frontend/src/pages/PostWritePage.js
@@ -145,6 +145,7 @@ const PostWritePage = ({ history, location }) => {
   const handleSubmitBtn = async (data) => {
     if (!loading) {
       setLoading((loading) => !loading);
+      setLoadingText('게시글 저장중..');
       if (store) {
         try {
           await SaveStore({ ...data, images: files });
@@ -191,7 +192,7 @@ const PostWritePage = ({ history, location }) => {
     <Body>
       <Header />
       <main>
-        {/* <AntModal visible={loading} loadingText={loadingText} /> */}
+        {loading && <AntModal visible="true" loadingText={loadingText} />}
         <StyledForm onSubmit={handleSubmit(handleSubmitBtn)}>
           <div className="write_page_header">
             <h1>리뷰 작성</h1>
